Let application set group-/pairwise-mode on request & add JUnit tests
Implemented functionality to allow the application to set information
related to pairwise or group mode on a request. This is done by encoding
information in the OSCORE option. I also added a number of JUnit tests
for instance for decryption and encryption verified against known values.
Furthermore I fixed an issue about the external AAD used during encryption
and decryption in Group OSCORE pairwise mode.