Add key derivation JUnit tests based on Group OSCORE interop test spec
I have now started adding JUnit tests for key derivation based on some
values used in the Group OSCORE interop test spec.

This can be used in the upcoming interop tests during the IETF 108
Hackathon to check that we derive the same keys.

The repository with the interop test specification:
https://github.com/EricssonResearch/Multicast-OSCOAP