Signature verification before decryption & small fix in COSE for this
Changed the order of decryption and signature verification, so that
signature verification comes first (based on discussions in the recent
meetings about Group OSCORE).

There was also a need for small fix in COSE to ensure that the element
protected in the ToBeSigned structure is an empty CBOR byte string
rather than a CBOR NULL value. This is similar to the issue I discovered
during interop tests with Jim last year. Then I fixed it in the method
for calculating the AAD, but that relied on doing decryption before
signature verification. Which is due to the method for calculating the
AAD not being called unless the object is decrypted.