Use multicast handler for multicast requests from proxy & related fixes
Worked with the functionality for proxy multicast support. I did the following fixes:
- Make the proxy send multicast requests with a multicast handler
- Ensure support for both Proxy-URI and Proxy-Scheme+Uri-* options with multicast in the proxy
- Additional types of test requests via the proxy added in ExampleProxy2CoapClient
- Listen for incoming multicast requests in the ExampleCoapServer
    - Listen on one port for multicast requests but respond unicast from different port
- Some code restructuring

The current state for multicast support is that the proxy fails with an
ExchangeCompleteException when relaying the response from the second server.
This is likely because this exchange has already been responded to and the
corresponding request from the client is a unicast request, so multiple
responses are not expected for the exchange. So functionality works for the
request(s) Client->Proxy->Servers, and then for the response(s) Servers->Proxy,
but fails there on the proxy.