Ensure that body_protected in COSE ToBeSigned is never null when signing
Previously there was a bug that caused body_protected to be null when
algorithms without TAG were used.

Algorithms which caused getAADBytes() to be called were not affected, as
that method was in fact performing steps on the COSE_Encrypt0 object
itself that eliminated the presence of the null. getAADBytes() is used
specifically for building the full AAD for encryption/decryption. Since
the algorithms without TAG never caused that method to be called, only
they were affected by this problem.